Add update deadline on access and remove expired options for LRU
