+++ pkg/cache/cache.go
@@ -15,7 +15,6 @@ package cache
 
 import (
 	"sync"
-	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
 
@@ -70,76 +69,5 @@ func (c *LRUCache[K, V]) Close() error {
 	defer c.mtx.Unlock()
 
 	c.lru.Purge()
-	return c.lru.Close()
-}
-
-type LRUCacheWithTTL[K comparable, V any] struct {
-	lru *lru.LRU[K, valueWithDeadline[V]]
-	mtx *sync.RWMutex
-
-	ttl time.Duration
-}
-
-type valueWithDeadline[V any] struct {
-	value    V
-	deadline time.Time
-}
-
-func NewLRUCacheWithTTL[K comparable, V any](reg prometheus.Registerer, maxEntries int, ttl time.Duration) *LRUCacheWithTTL[K, V] {
-	return &LRUCacheWithTTL[K, V]{
-		lru: lru.New[K, valueWithDeadline[V]](reg, lru.WithMaxSize[K, valueWithDeadline[V]](maxEntries)),
-		mtx: &sync.RWMutex{},
-		ttl: ttl,
-	}
-}
-
-func (c *LRUCacheWithTTL[K, V]) Add(key K, value V) {
-	c.mtx.Lock()
-	defer c.mtx.Unlock()
-	c.lru.Add(key, valueWithDeadline[V]{
-		value:    value,
-		deadline: time.Now().Add(c.ttl),
-	})
-}
-
-func (c *LRUCacheWithTTL[K, V]) Get(key K) (V, bool) {
-	c.mtx.RLock()
-	v, ok := c.lru.Get(key)
-	c.mtx.RUnlock()
-	if !ok {
-		return v.value, false
-	}
-	if v.deadline.Before(time.Now()) {
-		c.mtx.Lock()
-		c.lru.Remove(key)
-		c.mtx.Unlock()
-		return v.value, false
-	}
-	return v.value, true
-}
-
-func (c *LRUCacheWithTTL[K, V]) Peek(key K) (V, bool) {
-	c.mtx.RLock()
-	defer c.mtx.RUnlock()
-	v, ok := c.lru.Peek(key)
-	return v.value, ok
-}
-
-func (c *LRUCacheWithTTL[K, V]) Remove(key K) {
-	c.mtx.Lock()
-	defer c.mtx.Unlock()
-	c.lru.Remove(key)
-}
-
-func (c *LRUCacheWithTTL[K, V]) Purge() {
-	c.mtx.Lock()
-	defer c.mtx.Unlock()
-	c.lru.Purge()
-}
-
-func (c *LRUCacheWithTTL[K, V]) Close() error {
-	c.mtx.Lock()
-	defer c.mtx.Unlock()
-
 	return c.lru.Close()
 }
